refactor: rename coin adapter classes
